+++ src/main/java/oshi/software/os/linux/LinuxHardwareAbstractionLayer.java
@@ -24,8 +24,10 @@ import oshi.hardware.HardwareAbstractionLayer;
 import oshi.hardware.Memory;
 import oshi.hardware.PowerSource;
 import oshi.hardware.Processor;
+import oshi.software.os.OSFileStore;
 import oshi.software.os.linux.proc.CentralProcessor;
 import oshi.software.os.linux.proc.GlobalMemory;
+import oshi.software.os.linux.proc.LinuxFileSystem;
 import oshi.software.os.linux.proc.LinuxPowerSource;
 import oshi.util.FileUtil;
 
@@ -117,6 +119,10 @@ public class LinuxHardwareAbstractionLayer implements HardwareAbstractionLayer {
 
 	public PowerSource[] getPowerSources() {
 		return LinuxPowerSource.getPowerSources();
+	}
+
+	public OSFileStore[] getFileStores() {
+		return LinuxFileSystem.getFileStores();
 	}
 
 }